run inference on polygons
this script will read a testing folder containing all the json file and the map tif file. It will patchify the big map & cut legend based on the coordinate given in the json file, and save these patches in a different subfolder. Then for each patch image, it will start the inference by loading a pretrained model, get a prediction, and then unpatchify back to a big image.